--- EnemySpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2df6699c40b275741a3816f867e843f6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
